fix(qwik-city): duplicate ServerError class during dev
