+++ .changeset/tired-spies-call.md
@@ -1,0 +1,5 @@
+---
+'@builder.io/qwik-city': patch
+---
+
+FIX: Duplicate ServerError class during dev mode
