+++ packages/docs/src/routes/api/qwik-city-middleware-request-handler/index.mdx
@@ -1659,7 +1659,7 @@ number
 </td></tr>
 </tbody></table>
 
-[Edit this section](https://github.com/QwikDev/qwik/tree/main/packages/qwik-city/src/middleware/request-handler/error-handler.ts)
+[Edit this section](https://github.com/QwikDev/qwik/tree/main/packages/qwik-city/src/middleware/request-handler/server-error.ts)
 
 ## ServerRenderOptions
