+++ packages/qwik-city/src/middleware/request-handler/error-handler.ts
@@ -1,24 +1,4 @@
 /** @public */
-export class ServerError<T = any> extends Error {
-  constructor(
-    public status: number,
-    public data: T
-  ) {
-    super(typeof data === 'string' ? data : undefined);
-  }
-}
-
-/** @deprecated */
-export class ErrorResponse extends ServerError {
-  constructor(
-    public status: number,
-    message?: string
-  ) {
-    super(status, message);
-  }
-}
-
-/** @public */
 export function getErrorHtml(status: number, e: any) {
   let message = 'Server Error';
