+++ packages/qwik-city/src/middleware/request-handler/index.ts
@@ -1,5 +1,6 @@
-export { getErrorHtml, ServerError } from './error-handler';
+export { getErrorHtml } from './error-handler';
 export { mergeHeadersCookies } from './cookie';
+export { ServerError } from './server-error';
 export { AbortMessage, RedirectMessage } from './redirect-handler';
 export { RewriteMessage } from './rewrite-handler';
 export { requestHandler } from './request-handler';
