+++ packages/qwik-city/src/middleware/request-handler/types.ts
@@ -1,7 +1,7 @@
 import type { _deserializeData, _serializeData, _verifySerializable } from '@builder.io/qwik';
 import type { Action, FailReturn, Loader, QwikCityPlan } from '@builder.io/qwik-city';
 import type { Render, RenderOptions } from '@builder.io/qwik/server';
-import type { ServerError } from './error-handler';
+import type { ServerError } from './server-error';
 import type { AbortMessage, RedirectMessage } from './redirect-handler';
 import type { RequestEventInternal } from './request-event';
 import type { RewriteMessage } from './rewrite-handler';
