+++ packages/qwik-city/src/middleware/request-handler/user-response.ts
@@ -1,7 +1,6 @@
 import type { RequestEvent, RequestHandler } from '@builder.io/qwik-city';
-import type { RebuildRouteInfoInternal, LoadedRoute } from '../../runtime/src/types';
-import { ServerError, getErrorHtml, minimalHtmlResponse } from './error-handler';
-import { AbortMessage, RedirectMessage } from './redirect-handler';
+import type { LoadedRoute, RebuildRouteInfoInternal } from '../../runtime/src/types';
+import { getErrorHtml } from './error-handler';
 import {
   RequestEvQwikSerializer,
   createRequestEvent,
@@ -10,7 +9,13 @@ import {
 } from './request-event';
 import { encoder } from './resolve-request-handlers';
 import type { QwikSerializer, ServerRequestEvent, StatusCodes } from './types';
-import { RewriteMessage } from './rewrite-handler';
+// Import separately to avoid duplicate imports in the vite dev server
+import {
+  AbortMessage,
+  RedirectMessage,
+  ServerError,
+  RewriteMessage,
+} from '@builder.io/qwik-city/middleware/request-handler';
 
 export interface QwikCityRun<T> {
   response: Promise<T | null>;
@@ -112,7 +117,7 @@ async function runNext(
             const stream = requestEv.getWritableStream();
             if (!stream.locked) {
               const writer = stream.getWriter();
-              await writer.write(encoder.encode(minimalHtmlResponse(500, 'Internal Server Error')));
+              await writer.write(encoder.encode(getErrorHtml(500, 'Internal Server Error')));
               await writer.close();
             }
           } catch {
